# poxa_back/docs/swagger.yaml
definitions:
  models.Post:
    properties:
      content:
        type: string
      cover_image:
        type: string
      id:
        type: integer
      tags:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
paths:
  /posts:
    get:
      description: Retorna todos os posts salvos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      summary: Lista todos os posts
      tags:
      - Posts
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cria um novo post
      tags:
      - Posts
  /posts/{id}:
    delete:
      parameters:
      - description: ID do post
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deleta um post
      tags:
      - Posts
    get:
      parameters:
      - description: ID do post
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retorna um post por ID
      tags:
      - Posts
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do post
        in: path
        name: id
        required: true
        type: integer
      - description: Novos dados do post
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualiza um post existente
      tags:
      - Posts
swagger: "2.0"